back-end/requirements.txt
fastapi
uvicorn
langchain>=0.1.0
langchain-community>=0.0.10
openai
pymupdf
python-dotenv
pandas>=2.0.0
pydantic
requests
tiktoken==0.9.0
numpy>=1.24.0
faker>=18.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
pyarrow>=14.0.0 
pymupdf
apscheduler>=3.10.4
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
grafana-api==1.0.3
docker>=7.0.0
python-multipart>=0.0.6